Refactor PictureDocket into Composer component
